# backend/docs/swagger.json
{
    "components": {
        "schemas": {
            "APIError": {
                "properties": {
                    "errors": {
                        "description": "detalle de validación, uno por campo",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "message": {
                        "example": "Recurso no encontrado",
                        "type": "string"
                    },
                    "path": {
                        "example": "/api/v1/meter/getById/M-999",
                        "type": "string"
                    },
                    "status_code": {
                        "example": 404,
                        "type": "integer"
                    },
                    "timestamp": {
                        "example": "2026-09-24T15:00:00Z",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AnalysisRun": {
                "properties": {
                    "duration_ms": {
                        "example": 38,
                        "type": "integer"
                    },
                    "error": {
                        "type": "string"
                    },
                    "finished_at": {
                        "description": "null mientras corre",
                        "type": "string"
                    },
                    "id": {
                        "example": "AN-0002",
                        "type": "string"
                    },
                    "narrator": {
                        "description": "quién redacta las explicaciones",
                        "example": "openai:gpt-4o-mini",
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/RunStatus"
                    },
                    "steps": {
                        "items": {
                            "$ref": "#/components/schemas/AnalysisStep"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "summary": {
                        "$ref": "#/components/schemas/AnalysisSummary",
                        "description": "null mientras corre o si falló"
                    },
                    "trigger": {
                        "$ref": "#/components/schemas/Trigger"
                    }
                },
                "type": "object"
            },
            "AnalysisStep": {
                "properties": {
                    "detail": {
                        "example": "3 cambios sostenidos · 1 medidor con lecturas incoherentes",
                        "type": "string"
                    },
                    "key": {
                        "example": "detection",
                        "type": "string"
                    },
                    "label": {
                        "example": "Detección",
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/StepStatus"
                    }
                },
                "type": "object"
            },
            "AnalysisSummary": {
                "properties": {
                    "anomalies_detected": {
                        "example": 4,
                        "type": "integer"
                    },
                    "avg_confidence": {
                        "example": 0.9,
                        "type": "number"
                    },
                    "by_type": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object"
                    },
                    "requiring_attention": {
                        "description": "severidad HIGH",
                        "example": 2,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Anomaly": {
                "properties": {
                    "anomaly": {
                        "type": "boolean"
                    },
                    "confidence": {
                        "type": "number"
                    },
                    "detected_at": {
                        "type": "string"
                    },
                    "evidence": {
                        "$ref": "#/components/schemas/Evidence"
                    },
                    "meter_id": {
                        "type": "string"
                    },
                    "narrated_by": {
                        "description": "quién redactó reason y recommended_action: \"engine\" u \"openai:\u003cmodelo\u003e\"",
                        "example": "engine",
                        "type": "string"
                    },
                    "priority": {
                        "description": "1 = revisar primero",
                        "type": "integer"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "recommended_action": {
                        "type": "string"
                    },
                    "severity": {
                        "$ref": "#/components/schemas/Severity"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AnomalyType"
                    }
                },
                "type": "object"
            },
            "AnomalyBrief": {
                "properties": {
                    "anomaly": {
                        "description": "false en los falsos positivos",
                        "example": true,
                        "type": "boolean"
                    },
                    "confidence": {
                        "example": 0.95,
                        "type": "number"
                    },
                    "detected_at": {
                        "type": "string"
                    },
                    "meter_id": {
                        "example": "M-109",
                        "type": "string"
                    },
                    "priority": {
                        "example": 1,
                        "type": "integer"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "recommended_action": {
                        "type": "string"
                    },
                    "severity": {
                        "$ref": "#/components/schemas/Severity"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AnomalyType"
                    }
                },
                "type": "object"
            },
            "AnomalyType": {
                "enum": [
                    "REAL_ANOMALY",
                    "EXPLAINABLE_ANOMALY",
                    "FALSE_POSITIVE",
                    "DATA_QUALITY"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "TypeReal",
                    "TypeExplainable",
                    "TypeFalsePositive",
                    "TypeDataQuality"
                ]
            },
...
            "Consumption": {
                "properties": {
                    "baseline_daily_kwh": {
                        "description": "suma de los baselines diarios",
                        "example": 12980.3,
                        "type": "number"
                    },
                    "last_day_kwh": {
                        "description": "últimas 24 h",
                        "example": 14210.5,
                        "type": "number"
                    },
                    "period_kwh": {
                        "description": "total de los 14 días",
                        "example": 167420.5,
                        "type": "number"
                    },
                    "variation_pct": {
                        "example": 9.48,
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
            "DashboardSummary": {
                "properties": {
                    "ai_confidence": {
                        "description": "confianza promedio de las anomalías",
                        "example": 0.9,
                        "type": "number"
                    },
                    "anomalies_by_type": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object"
                    },
                    "anomalies_detected": {
                        "description": "incluye falsos positivos",
                        "example": 4,
                        "type": "integer"
                    },
                    "consumption": {
                        "$ref": "#/components/schemas/Consumption"
                    },
                    "data_range": {
                        "$ref": "#/components/schemas/DataRange"
                    },
                    "last_analysis": {
                        "$ref": "#/components/schemas/LastAnalysis"
                    },
                    "priorities": {
                        "description": "ordenadas por prioridad",
                        "items": {
                            "$ref": "#/components/schemas/AnomalyBrief"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "requiring_attention": {
                        "description": "severidad HIGH",
                        "example": 2,
                        "type": "integer"
                    },
                    "status_counts": {
                        "$ref": "#/components/schemas/StatusCounts"
                    },
                    "total_meters": {
                        "example": 12,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "DataRange": {
                "properties": {
                    "from": {
                        "type": "string"
                    },
                    "to": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Evidence": {
                "properties": {
                    "baseline_daily_kwh": {
                        "type": "number"
                    },
                    "change": {
                        "$ref": "#/components/schemas/ConsumptionChange"
                    },
                    "checks": {
                        "items": {
                            "$ref": "#/components/schemas/Check"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "daily_change_pct": {
                        "type": "number"
                    },
                    "impact_kwh": {
                        "description": "energía de más (+) o de menos (-) frente a lo esperado",
                        "type": "number"
                    },
                    "last_day_kwh": {
                        "type": "number"
                    },
                    "quality": {
                        "$ref": "#/components/schemas/Report"
                    },
                    "related_events": {
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "variables": {
                        "items": {
                            "$ref": "#/components/schemas/VariableChange"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "window_end": {
                        "type": "string"
                    },
                    "window_start": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LastAnalysis": {
                "properties": {
                    "finished_at": {
                        "type": "string"
                    },
                    "id": {
                        "example": "AN-0002",
                        "type": "string"
                    },
                    "narrator": {
                        "example": "engine",
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/RunStatus"
                    },
                    "trigger": {
                        "$ref": "#/components/schemas/Trigger"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "properties": {
                    "email": {
                        "example": "admin@energia.local",
                        "type": "string"
                    },
                    "password": {
                        "example": "admin123",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "RunStatus": {
                "enum": [
                    "RUNNING",
                    "COMPLETED",
                    "FAILED"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "RunRunning",
                    "RunCompleted",
                    "RunFailed"
                ]
            },
            "Severity": {
                "enum": [
                    "HIGH",
                    "MEDIUM",
                    "LOW"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "SeverityHigh",
                    "SeverityMedium",
                    "SeverityLow"
                ]
            },
...
            "StatusCounts": {
                "properties": {
                    "alert": {
                        "example": 1,
                        "type": "integer"
                    },
                    "critical": {
                        "example": 2,
                        "type": "integer"
                    },
                    "normal": {
                        "example": 9,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "StepStatus": {
                "enum": [
                    "PENDING",
                    "RUNNING",
                    "DONE",
                    "FAILED"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "StepPending",
                    "StepRunning",
                    "StepDone",
                    "StepFailed"
                ]
            },
            "TokenResponse": {
                "properties": {
                    "access_token": {
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                        "type": "string"
                    },
                    "expires_in": {
                        "description": "segundos",
                        "example": 28800,
                        "type": "integer"
                    },
                    "token_type": {
                        "example": "Bearer",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Trigger": {
                "enum": [
                    "STARTUP",
                    "MANUAL"
                ],
                "type": "string",
                "x-enum-comments": {
                    "TriggerManual": "botón Run AI Analysis",
                    "TriggerStartup": "al arrancar la API"
                },
                "x-enum-varnames": [
                    "TriggerStartup",
                    "TriggerManual"
                ]
            },
            "User": {
                "properties": {
                    "email": {
                        "example": "admin@energia.local",
                        "type": "string"
                    },
                    "name": {
                        "example": "Operador",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "bearerFormat": "JWT",
                "description": "Pega solo el token (sin \"Bearer\").",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "API para monitorear medidores eléctricos y priorizar sus anomalías.\nPrimero haz login en POST /auth/login, copia el `access_token` y pégalo en **Authorize**.",
        "title": "EnergIA API",
        "version": "1.0"
    },
...
    "paths": {
        "/ai/analysis/latest": {
            "get": {
                "description": "El análisis más reciente, terminado o en curso.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisRun"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Último análisis",
                "tags": [
                    "AI"
                ]
            }
        },
        "/ai/analysis/{id}": {
            "get": {
                "description": "Pasos con su estado (PENDING, RUNNING, DONE, FAILED) y, al terminar, el resumen.",
                "parameters": [
                    {
                        "description": "ID del análisis",
                        "example": "AN-0002",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisRun"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Estado de un análisis",
                "tags": [
                    "AI"
                ]
            }
        },
        "/ai/analyze": {
            "post": {
                "description": "Lanza el análisis en segundo plano y responde 202 con el análisis en curso.\nPasos: Lecturas → Baseline → Detección → Correlación → Eventos → Explicación → Recomendación.\nConsulta el avance con GET /ai/analysis/{id}. Si ya hay uno corriendo, devuelve ese (200).",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisRun"
                                }
                            }
                        },
                        "description": "Ya había un análisis en curso"
                    },
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisRun"
                                }
                            }
                        },
                        "description": "Análisis iniciado"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Run AI Analysis",
                "tags": [
                    "AI"
                ]
            }
        },
        "/anomaly/getAll": {
            "get": {
                "description": "Todas las anomalías, ordenadas por prioridad (1 = revisar primero), con explicación,\nacción recomendada y evidencia. Los falsos positivos vienen con `anomaly: false`.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Anomaly"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Anomalías detectadas",
                "tags": [
                    "Anomalies"
                ]
            }
        },
        "/auth/login": {
            "post": {
                "description": "Valida email y contraseña y devuelve un JWT. Úsalo en el header `Authorization: Bearer \u003ctoken\u003e`.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest",
                                "summary": "body",
                                "description": "Credenciales"
                            }
                        }
                    },
                    "description": "Credenciales",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "Iniciar sesión",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/me": {
            "get": {
                "description": "Devuelve el usuario del token. Sirve al front para validar la sesión al recargar.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Usuario autenticado",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/dashboard/getSummary": {
            "get": {
                "description": "Medidores por estado, anomalías por tipo, consumo total vs. baseline y la lista de prioridades.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardSummary"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Resumen del dashboard",
                "tags": [
                    "Dashboard"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}